Move all extractors into their own package
