+++ olio/extractors/health_extractor.go
@@ -1,0 +1,13 @@
+package extractors
+
+import (
+	"time"
+)
+
+type UptimeExtractor interface {
+	GetUptime() time.Duration
+}
+
+type DbHealthExtractor interface {
+	GetDbExtractor() DbExtractor
+}
