+++ olio/extractors/version_extractor.go
@@ -1,0 +1,6 @@
+package extractors
+
+type VersionExtractor interface {
+	GetVersion() string
+	GetAppName() string
+}
