+++ olio/service/base.go
@@ -3,6 +3,7 @@ package service
 import (
 	"github.com/gin-gonic/gin"
 	network "github.com/glibs/gin-webserver"
+	"github.com/olioapps/service-skeleton-go/olio/extractors"
 	olioMiddleware "github.com/olioapps/service-skeleton-go/olio/service/middleware"
 	olioResources "github.com/olioapps/service-skeleton-go/olio/service/resources"
 	"github.com/olioapps/service-skeleton-go/olio/util"
@@ -66,15 +67,15 @@ func (obs *OlioBaseService) AddDaemon(daemon OlioDaemon) {
 	obs.daemons = append(obs.daemons, daemon)
 }
 
-func (obs *OlioBaseService) AddVersionExtractor(versionExtractor olioResources.VersionExtractor) {
+func (obs *OlioBaseService) AddVersionExtractor(versionExtractor extractors.VersionExtractor) {
 	obs.versionResource.AddVersionExtractor(versionExtractor)
 }
 
-func (obs *OlioBaseService) AddUptimeExtractor(uptimeExtractor olioResources.UptimeExtractor) {
+func (obs *OlioBaseService) AddUptimeExtractor(uptimeExtractor extractors.UptimeExtractor) {
 	obs.healthResource.AddUptimeExtractor(uptimeExtractor)
 }
 
-func (obs *OlioBaseService) AddDbHealthExtractor(dbHealthExtractor olioResources.DbHealthExtractor) {
+func (obs *OlioBaseService) AddDbHealthExtractor(dbHealthExtractor extractors.DbHealthExtractor) {
 	obs.healthResource.AddDbHealthExtractor(dbHealthExtractor)
 }
